+++ src/Factory/AeatResponseFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Flux\VerifactuBundle\Factory;
 
+use Flux\VerifactuBundle\Contract\AeatResponseInterface;
 use Flux\VerifactuBundle\Dto\AeatResponseDto;
 use Flux\VerifactuBundle\Transformer\AeatResponseTransformer;
 use Flux\VerifactuBundle\Validator\ContractsValidator;
@@ -17,11 +18,16 @@ final readonly class AeatResponseFactory
     ) {
     }
 
-    public function makeValidatedAeatResponseDtoFromModel(AeatResponse $model): AeatResponseDto
+    public function makeValidatedAeatResponseDtoFromModel(AeatResponse $model): AeatResponseInterface
     {
         $validatedAeatResponseDto = $this->aeatResponseTransformer->transformModelToDto($model);
         $this->validator->validate($validatedAeatResponseDto);
 
         return $validatedAeatResponseDto;
+    }
+
+    public function getJsonStringFromAeatResponseDto(AeatResponseDto $dto): string
+    {
+        return $this->aeatResponseTransformer->transformDtoToJson($dto);
     }
 }