+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 CHANGELOG
 =========
 
+0.1.5
+-----
+ 
+ * Add QrCode handler
+ * Add Json serialization for Aeat responses
+ * Add AeatResponseDto test
+ * Enable Symfony 8.0 compatibility
+
+0.1.4
+-----
+ 
+ * Add more DTO unit tests
+ * Add BreakdownDetail transformer
+ * Add AeatResponse factory
+
 0.1.3
 -----